requirements.txt
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
scipy>=1.7.0
statsmodels>=0.13.0
pandas-datareader>=0.10.0
cookiecutter>=1.7.0
ipykernel>=6.0.0
feature_engine>=1.1.0
mlflow>=1.20.0
click>=8.0.0
python-dotenv>=0.19.0
pyyaml>=5.4.0
fastapi>=0.95.0
uvicorn>=0.21.0

